requirements.txt
streamlit==1.17.0
easyocr==1.6.2
opencv-python-headless==4.5.5.64
numpy==1.21.2
Pillow==9.0.1
streamlit-image-zoom==0.1.0